tflite_requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
boto3==1.34.13
botocore==1.34.13
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.3
openai==0.28.0
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.1
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
tenacity==8.2.3
tflite-runtime==2.14.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0
